3/5/13 FIRST DRAFT DONE.
The complete first draft of the entire game is done!
This is such a happy moment, I've been working on this for what, two
weeks? Three? I don't remember.
Finished scene four.
Had quite a bit of trouble with talking to Elily in all the different
places and making it possible to get to the different date locations.
Had trouble with the elevator.
All of these issues were not with the code, but instead, when it was
implemented.
All is fixed.
Point system (which was originally only in part 1) is not in all four
parts.