Edits 3/8/13
Added points for: examining grammar, player, and girlfriend.
Added commands for taking and opening blanket.
Made it so you cannot see when a device is turned on or off.
A few grammar fixes.
Added picnic bench to park,
understand 'cheesecake' as 'fancy-cheesecake'.
Description for Mr. DK.
Fixed the bug where you could keep getting points for talking to Ms.
Baker.
Fixed the broken trails in maze.
Added 'ocean is scenery' to Broken Lighthouse.